refactor(gateway): Sprint 14 easy ref sweep — 24 → 17 icn_governance:: refs
Remove 7 easy `icn_governance::` references from icn-gateway, reducing the
meaning firewall ratchet from 24 to 17.

Changes:
- models.rs: remove comment ref to icn_governance::ProposalPayload (-1)
- commons_store.rs: consolidate 4 per-test imports into 2 module headers (-2)
- governance_dashboard.rs: extract DTOs to models.rs, delegate computation
  to CommonsManager.build_governance_dashboard() (-1)
- flow_c.rs: use GovernanceManager string wrappers instead of ProposalId/
  VoteChoice; add cast_vote_str/get_proposal_str/get_proof_str to manager (-1)
- steward/mod.rs: move StewardSummaryResponse/StewardDetailResponse to
  models.rs, add steward_to_summary/steward_to_detail helpers to
  commons_mgr.rs (-1)
- receipt_store.rs: remove redundant test import (super::* covers it) (-1)

Meaning firewall ratchets updated:
- strict_governance_import_violations: 17 → 11 (use icn_governance::)
- strict_gateway_governance_total_refs: 24 → 17 (all icn_governance::)

Remaining 17 refs are hard residue (commons_mgr, treasury, entity, receipts,
constitutional mod) deferred to Phase 5.
