+++ icn/crates/icn-gateway/src/api/flow_c.rs
@@ -15,7 +15,6 @@ use crate::middleware::{get_claims, require_scope};
 use crate::models::CastVoteRequest;
 use crate::notifications::NotificationService;
 use crate::treasury_mgr::GatewayTreasuryManager;
-use icn_governance::{ProposalId, VoteChoice};
 use icn_identity::Did;
 
 /// POST /v1/coops/{coop_id}/proposals - Propose treasury spend.
@@ -62,34 +61,27 @@ pub async fn cast_vote_alias(
 
     crate::validation::validate_vote_comment(&req.comment)?;
 
-    let choice = match req.choice.to_lowercase().as_str() {
-        "for" => VoteChoice::For,
-        "against" => VoteChoice::Against,
-        "abstain" => VoteChoice::Abstain,
-        _ => {
-            return Err(GatewayError::BadRequest(format!(
-                "Invalid vote choice: {}",
-                req.choice
-            )))
-        }
-    };
+    let proposal_id_str = id.into_inner();
 
-    let proposal_id = ProposalId(id.into_inner());
     gov_mgr
-        .cast_vote(
-            proposal_id.clone(),
+        .cast_vote_str(
+            proposal_id_str.clone(),
             voter_did.clone(),
-            choice,
+            &req.choice,
             req.comment.clone(),
         )
-        .await?;
+        .await
+        .map_err(|e| GatewayError::BadRequest(e.to_string()))?;
 
-    let proposal = gov_mgr.get_proposal(&proposal_id).await?.ok_or_else(|| {
-        GatewayError::InternalError("Vote cast succeeded but proposal not found".to_string())
-    })?;
+    let proposal = gov_mgr
+        .get_proposal_str(&proposal_id_str)
+        .await?
+        .ok_or_else(|| {
+            GatewayError::InternalError("Vote cast succeeded but proposal not found".to_string())
+        })?;
 
     let vote_event = GatewayEvent::GovernanceVoteCast {
-        proposal_id: proposal_id.0.clone(),
+        proposal_id: proposal_id_str.clone(),
         domain_id: proposal.domain_id.0.clone(),
         voter: voter_did.to_string(),
         choice: req.choice.clone(),
@@ -117,31 +109,32 @@ pub async fn get_proof_alias(
 ) -> Result<HttpResponse> {
     require_scope(&http_req, "governance:read")?;
 
-    let proposal_id = ProposalId(id.into_inner());
+    let proposal_id_str = id.into_inner();
 
-    let proof = gov_mgr.get_proof(&proposal_id).await?.ok_or_else(|| {
-        GatewayError::NotFound(format!("No proof found for proposal: {}", proposal_id.0))
-    })?;
+    let proof = gov_mgr
+        .get_proof_str(&proposal_id_str)
+        .await?
+        .ok_or_else(|| {
+            GatewayError::NotFound(format!("No proof found for proposal: {proposal_id_str}"))
+        })?;
 
     if !proof.verify_receipt() {
         tracing::warn!(
             "Invalid proof binding for proposal {} — not serving",
-            proposal_id.0
+            proposal_id_str
         );
         return Err(GatewayError::NotFound(format!(
-            "No valid proof found for proposal: {}",
-            proposal_id.0
+            "No valid proof found for proposal: {proposal_id_str}"
         )));
     }
 
     if proof.attestations.is_empty() {
         tracing::warn!(
             "Proof for proposal {} has no attestations — not serving",
-            proposal_id.0
+            proposal_id_str
         );
         return Err(GatewayError::NotFound(format!(
-            "No valid proof found for proposal: {}",
-            proposal_id.0
+            "No valid proof found for proposal: {proposal_id_str}"
         )));
     }
